requirements.txt
Django==3.2
celery==5.2.1
redis==4.0.2
django-celery-beat==2.2.1
requests==2.26.0
djangorestframework==3.12.4